게시글과 댓글 수정/삭제권한 추가

+++ app/controllers/wall_controller.rb
@@ -18,6 +18,7 @@ class WallController < ApplicationController
 
   def show_detail
     @post = Post.find(params[:id])
+    @comment_writer = User.where(id: session[:user_id])[0]
   end
 
   def write
@@ -26,6 +27,7 @@ class WallController < ApplicationController
 
   def write_complete
     post = Post.new
+    post.user_id = session[:user_id]
     post.category = params[:post_category]
     post.title = params[:post_title]
     post.content = params[:post_content]
@@ -41,6 +43,7 @@ class WallController < ApplicationController
   
   def write_comment_complete
     comment = Comment.new
+    comment.user_id = session[:user_id]
     comment.post_id = params[:post_id]
     comment.content = params[:comment_content]
     comment.save
@@ -51,13 +54,22 @@ class WallController < ApplicationController
   
   def delete_comment_complete
     comment = Comment.find(params[:id])
-    comment.destroy
-    flash[:alert] = "댓글이 삭제되었습니다."
-    redirect_to "/wall/show_detail/#{comment.post_id}"
+    if comment.user_id == session[:user_id]
+      comment.destroy
+      flash[:alert] = "댓글이 삭제되었습니다."
+      redirect_to "/wall/show_detail/#{comment.post_id}"
+    else
+      flash[:alert] = "해당 댓글의 삭제 권한이 없습니다."
+      redirect_to :back
+    end
   end
 
   def edit
     @post = Post.find(params[:id])
+    if @post.user_id != session[:user_id]
+      flash[:alert] = "수정 권한이 없습니다."
+      redirect_to :back
+    end
   end
 
   def edit_complete
@@ -78,8 +90,13 @@ class WallController < ApplicationController
 
   def delete_complete
     post = Post.find(params[:id])
-    post.destroy
-    flash[:alert] = "삭제되었습니다."
-    redirect_to "/"
+    if post.user_id == session[:user_id]
+      post.destroy
+      flash[:alert] = "삭제되었습니다."
+      redirect_to "/"
+    else
+      flash[:alert] = "삭제 권한이 없습니다."
+      redirect_to :back
+    end
   end
 end
